+++ packages/terra-graphs-docs/CHANGELOG.md
@@ -25,3 +25,4 @@
   * Updated Bubble graph documentation and examples to be carbon centric.
   * Updated Line graph documentation and examples to be carbon-centric.
   * Updated Combination graph documentation and examples to be carbon-centric.
+  * Updated Gantt chart documentation and examples to be carbon-centric